requirements.txt
requests>=2.28.0
tkinter
pillow>=9.0.0
python-dateutil>=2.8.2